+++ src/org/elixir_lang/beam/chunk/debug_info/v1/erl_abstract_code/abstract_code_compiler_options/abstract_code/Call.kt
@@ -1,13 +1,17 @@
 package org.elixir_lang.beam.chunk.debug_info.v1.erl_abstract_code.abstract_code_compiler_options.abstract_code
 
-import com.ericsson.otp.erlang.OtpErlangAtom
 import com.ericsson.otp.erlang.OtpErlangList
 import com.ericsson.otp.erlang.OtpErlangObject
 import com.ericsson.otp.erlang.OtpErlangTuple
+import org.elixir_lang.NameArity
+import org.elixir_lang.beam.Decompiler
 import org.elixir_lang.beam.chunk.debug_info.v1.erl_abstract_code.abstract_code_compiler_options.AbstractCode
 import org.elixir_lang.beam.chunk.debug_info.v1.erl_abstract_code.abstract_code_compiler_options.AbstractCode.ifTag
+import org.elixir_lang.beam.decompiler.InfixOperator
+import org.elixir_lang.beam.decompiler.PrefixOperator
+import org.elixir_lang.beam.decompiler.Unquoted
 import org.elixir_lang.code.Identifier.inspectAsFunction
-
+import org.elixir_lang.toOtpErlangList
 
 object Call {
     fun ifToMacroStringDeclaredScope(term: OtpErlangObject): MacroStringDeclaredScope? =
@@ -36,51 +40,66 @@ object Call {
     private fun argumentsToMacroString(term: OtpErlangList): String =
             term.joinToString(", ") { AbstractCode.toMacroStringDeclaredScope(it, Scope.EMPTY).macroString }
 
-    private fun argumentsToMacroString(term: OtpErlangObject): String =
+    private fun argumentsToMacroString(term: OtpErlangObject?): String =
         when (term) {
             is OtpErlangList -> argumentsToMacroString(term)
             else -> "unknown_arguments"
         }
 
     private fun localNamedFunctionCallToMacroString(
-            term: OtpErlangTuple,
-            argumentsMacroString: MacroString
-    ): MacroString {
-        val nameMacroString = nameMacroString(term)
+            name: OtpErlangObject?,
+            arguments: OtpErlangObject?
+    ): MacroString =
+        if (name != null) {
+            val elixirAtom = Atom.toElixirAtom(name)
 
-        return "$nameMacroString($argumentsMacroString)"
-    }
+            if (elixirAtom != null) {
+                val function = inspectAsFunction(elixirAtom.atomValue(), local = true)
+                val argumentList = arguments?.toOtpErlangList()
+                val argumentsMacroString = argumentsToMacroString(arguments)
+
+                if (argumentList != null) {
+                    val nameArity = NameArity(elixirAtom.atomValue(), argumentList.arity())
+
+                    when (Decompiler.decompiler("erlang", nameArity)) {
+                        is PrefixOperator, is InfixOperator, is Unquoted -> {
+                            "apply(__MODULE__, :${function}, [${argumentsMacroString}])"
+                        }
+                        else -> {
+                            "${function}(${argumentsMacroString})"
+                        }
+                    }
+                } else {
+                    "${function}(${argumentsMacroString})"
+                }
+            } else {
+                val nameMacroString = AbstractCode.toMacroStringDeclaredScope(name, Scope.EMPTY).macroString
+                val argumentMacroString = argumentsToMacroString(arguments)
+
+                "${nameMacroString}(${argumentMacroString})"
+            }
+        } else {
+            val nameMacroString = "unknown_function"
+            val argumentMacroString = argumentsToMacroString(arguments)
+
+            "${nameMacroString}(${argumentMacroString})"
+        }
 
     private fun namedFunctionCallToMacroString(term: OtpErlangTuple): MacroString {
         val name = toName(term)
-        val argumentsMacroString = argumentsMacroString(term)
+        val arguments = toArguments(term)
 
         return Remote.ifTo(name) { remoteName ->
-            remoteNamedFunctionCallToMacroString(remoteName, argumentsMacroString)
-        } ?: localNamedFunctionCallToMacroString(term, argumentsMacroString)
+            remoteNamedFunctionCallToMacroString(remoteName, arguments)
+        } ?: localNamedFunctionCallToMacroString(name, arguments)
     }
-
-    private fun nameMacroString(term: OtpErlangTuple): String =
-            toName(term)
-                    ?.let { nameToMacroString(it) }
-                    ?: "missing_name"
-
-    private fun nameToMacroString(term: OtpErlangObject): String =
-            Atom.ifTo(term) {
-                Atom.toAtom(it)?.let { atom ->
-                    if (atom is OtpErlangAtom) {
-                        inspectAsFunction(atom, local = true)
-                    } else {
-                        null
-                    }
-                } ?: "call_unknown_function"
-            } ?: AbstractCode.toMacroStringDeclaredScope(term, Scope.EMPTY).macroString
 
     private fun remoteNamedFunctionCallToMacroString(
             remoteName: OtpErlangTuple,
-            argumentsMacroString: MacroString
+            arguments: OtpErlangObject?
     ): MacroString {
         val remoteFunction = Remote.toFunction(remoteName)
+        val argumentsMacroString = argumentsToMacroString(arguments)
 
         return if (Var.`is`(remoteFunction)) {
             val remoteModuleMacroString = Remote.moduleMacroString(remoteName)