+++ src/org/elixir_lang/psi/impl/QualifiableAliasImpl.kt
@@ -136,7 +136,7 @@ object QualifiableAliasImpl {
             is Operation,
             is QuotableKeywordPair,
             // containers
-            is ElixirList, is ElixirStructOperation, is ElixirTuple,
+            is ElixirContainerAssociationOperation, is ElixirList, is ElixirStructOperation, is ElixirTuple,
             // Top of file
             is ElixirFile,
             // Typing an alias on a new line in the body of function